+++ source/ui/ui_manager_load.cpp
@@ -54,18 +54,18 @@ std::uint32_t cUIManager::load(const std::string &_fileName)
             m_menu[m].scale = glm::vec2(m_menu[m].size.x, m_menu[m].size.y);
 
             // Calculate menu size in pixels
-            std::uint32_t menuSizePixels_x = (m_menu[m].size.x * m_window_w);
-            std::uint32_t menuSizePixels_y = (m_menu[m].size.y * m_window_h);
+            std::uint32_t menuSizePixels_x = static_cast<std::uint32_t>(m_menu[m].size.x * static_cast<float>(m_window_w));
+            std::uint32_t menuSizePixels_y = static_cast<std::uint32_t>(m_menu[m].size.y * static_cast<float>(m_window_h));
 
             // Calculate menu position in pixels
-            std::uint32_t menuPosPixels_x = (1 - m_menu[m].position.x) * m_window_w / 2;
-            std::uint32_t menuPosPixels_y = (1 - m_menu[m].position.y) * m_window_h / 2;
+            std::uint32_t menuPosPixels_x = static_cast<std::uint32_t>((1 - m_menu[m].position.x) * static_cast<float>(m_window_w) / 2.0f);
+            std::uint32_t menuPosPixels_y = static_cast<std::uint32_t>((1 - m_menu[m].position.y) * static_cast<float>(m_window_h) / 2.0f);
 
             // Calculate menu bounds
-            m_menu[m].positionMin.x = menuPosPixels_x - (menuSizePixels_x / 2);
-            m_menu[m].positionMin.y = menuPosPixels_y - (menuSizePixels_y / 2);
-            m_menu[m].positionMax.x = menuPosPixels_x + (menuSizePixels_x / 2);
-            m_menu[m].positionMax.y = menuPosPixels_y + (menuSizePixels_y / 2);
+            m_menu[m].positionMin.x = static_cast<float>(menuPosPixels_x - (menuSizePixels_x / 2));
+            m_menu[m].positionMin.y = static_cast<float>(menuPosPixels_y - (menuSizePixels_y / 2));
+            m_menu[m].positionMax.x = static_cast<float>(menuPosPixels_x + (menuSizePixels_x / 2));
+            m_menu[m].positionMax.y = static_cast<float>(menuPosPixels_y + (menuSizePixels_y / 2));
 
             // get a count of ui components for the current menu
             m_menu[m].numComponent = xmlUiFile.getInteger("<menu_component_count>", m + 1);
@@ -192,18 +192,18 @@ std::uint32_t cUIManager::load(const std::string &_fileName)
                     m_menu[m].component[c].scale = glm::vec2(m_menu[m].component[c].size.x, m_menu[m].component[c].size.y);
 
                     // Calculate menu component size in pixels
-                    std::uint32_t menuCompSizePixels_x = (m_menu[m].component[c].size.x * m_window_w);
-                    std::uint32_t menuCompSizePixels_y = (m_menu[m].component[c].size.y * m_window_h);
+                    std::uint32_t menuCompSizePixels_x = static_cast<uint32_t>(m_menu[m].component[c].size.x * static_cast<float>(m_window_w));
+                    std::uint32_t menuCompSizePixels_y = static_cast<uint32_t>(m_menu[m].component[c].size.y * static_cast<float>(m_window_h));
 
                     // Calculate menu component position in pixels
-                    std::uint32_t menuCompPosPixels_x = (1 + (m_menu[m].component[c].position.x)) * m_window_w / 2;
-                    std::uint32_t menuCompPosPixels_y = (1 - (m_menu[m].component[c].position.y)) * m_window_h / 2;
+                    std::uint32_t menuCompPosPixels_x = static_cast<uint32_t>((1 + (m_menu[m].component[c].position.x)) * static_cast<float>(m_window_w) / 2.0f);
+                    std::uint32_t menuCompPosPixels_y = static_cast<uint32_t>((1 - (m_menu[m].component[c].position.y)) * static_cast<float>(m_window_h) / 2.0f);
 
                     // Calculate menu component bounds
-                    m_menu[m].component[c].positionMin.x = menuCompPosPixels_x - (menuCompSizePixels_x / 2);
-                    m_menu[m].component[c].positionMin.y = menuCompPosPixels_y - (menuCompSizePixels_y / 2);
-                    m_menu[m].component[c].positionMax.x = menuCompPosPixels_x + (menuCompSizePixels_x / 2);
-                    m_menu[m].component[c].positionMax.y = menuCompPosPixels_y + (menuCompSizePixels_y / 2);
+                    m_menu[m].component[c].positionMin.x = static_cast<float>(menuCompPosPixels_x - (menuCompSizePixels_x / 2));
+                    m_menu[m].component[c].positionMin.y = static_cast<float>(menuCompPosPixels_y - (menuCompSizePixels_y / 2));
+                    m_menu[m].component[c].positionMax.x = static_cast<float>(menuCompPosPixels_x + (menuCompSizePixels_x / 2));
+                    m_menu[m].component[c].positionMax.y = static_cast<float>(menuCompPosPixels_y + (menuCompSizePixels_y / 2));
                 }
             }
